Move manifest-helpers.js
