+++ config/manifest-helpers.js
